RF: add json-file reading
